--- feature/Screenshots/Sources/Screenshots/CoursesOverviewView+Preview.swift
@@ -1,30 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Nityananda Zbil on 24.09.23.
-//
-
-import Common
-import Dependencies
-import SharedModels
-import SharedServices
-import SwiftUI
-@testable import Dashboard
-
-struct CoursesOverviewView_Previews: PreviewProvider {
-
-    static var previews: some View {
-        NavigationStack {
-            withDependencies { values in
-                values.courseService = CourseServiceStub()
-            } operation: {
-                CoursesOverviewView(viewModel: withDependencies({ values in
-                    values.courseService = CourseServiceStub()
-                }, operation: {
-                    CoursesOverviewViewModel()
-                }))
-            }
-        }
-    }
-}
